Sweets/feature/solidity prettier

+++ contracts/mocks/MockERC721.sol
@@ -7,18 +7,14 @@ import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract MockERC721 is ERC721, Ownable {
-
   uint256 public totalSupply;
 
   /// ============ Constructor ============
 
-  constructor(
-    string memory name,
-    string memory symbol
-  ) ERC721 (name, symbol) {}
+  constructor(string memory name, string memory symbol) ERC721(name, symbol) {}
 
-  function mintNft(uint numberOfTokens) public payable {
-    for(uint256 i = 0; i < numberOfTokens; i++) {
+  function mintNft(uint256 numberOfTokens) public payable {
+    for (uint256 i = 0; i < numberOfTokens; i++) {
       _safeMint(msg.sender, totalSupply++);
     }
   }
